+++ src/core/TransformFrames.ts
@@ -1,137 +1,89 @@
-import PropertyContainer from './PropertyContainer';
-import PropertyFactory from './PropertyFactory';
-import {
-  degToRads,
-  defaultVector,
-  initialDefaultFrame,
-} from './constant/index';
-import BaseLottieLayer from './BaseLottieLayer';
+import { degToRads, initialDefaultFrame } from './contant';
+import BaseLottieLayer from './layer/BaseLottieLayer';
+import ValueProperty from './property/ValueProperty';
+import MultiDimensionalProperty from './property/MultiDimensionalProperty';
+import KeyframedValueProperty from './property/KeyframedValueProperty';
+import KeyframedMultidimensionalProperty from './property/KeyframedMultidimensionalProperty';
 
-type Anchor = {
+type TypeMultiDimensionalProperty = {
   a: number;
-  ix: number;
   k: number[];
 }
 
-type Opacity = {
+type TypeValueProperty = {
   a: number;
-  ix: number;
   k: number;
 }
 
-type Rotation = {
-  a: number;
-  ix: number;
-  k: number;
-}
-
-type Scale = {
-  a: number;
-  ix: number;
-  k: number[];
+type KeyFrames = {
+  a: TypeMultiDimensionalProperty;
+  o: TypeValueProperty;
+  p: TypeMultiDimensionalProperty;
+  r: TypeValueProperty;
+  s: TypeMultiDimensionalProperty;
 }
 
 /**
  * transform property origin from tr or ks
- * @private
  */
-export default class TransformFrames extends PropertyContainer {
-  p;
-  px;
-  py;
-  pz;
-  rx;
-  ry;
-  rz;
-  or;
-  r;
-  sk;
-  sa;
-  a;
-  s;
-  o;
-  frameId: number = -1;
-  propType: string = 'transform';
-  private autoOriented: boolean = false; 
-  private orientation: number = 0; 
+export default class TransformFrames {
+  p: KeyframedMultidimensionalProperty;
+  r: KeyframedValueProperty;
+  a: KeyframedMultidimensionalProperty;
+  s: KeyframedMultidimensionalProperty;
+  o: KeyframedValueProperty;
+  private properties = [];
+  private autoOrient: boolean = false;
 
-  /**
-   * constructor about transform property
-   * @param {*} elem element node
-   * @param {*} data multidimensional value property data
-   */
-  constructor(elem, data) {
-    super(elem);
-    console.log('TransformFrames', elem, data)
+  static create(data, type, mult?): ValueProperty | MultiDimensionalProperty | KeyframedValueProperty | KeyframedMultidimensionalProperty {
+    if (!data.k.length) {
+      return new ValueProperty(data, mult);
+    }
+    else if (typeof (data.k[0]) === 'number') {
+      return new MultiDimensionalProperty(data, mult);
+    }
+    else {
+      if (type) {
+        return new KeyframedMultidimensionalProperty(data, mult);
+      }
+      else {
+        return new KeyframedValueProperty(data, mult);
+      }
+    }
+  }
 
-    this.p = PropertyFactory.create(data.p || { k: [0, 0, 0] }, 1, 0, this);
-    this.r = PropertyFactory.create(data.r || { k: 0 }, 0, degToRads, this);
-    this.a = PropertyFactory.create(data.a || { k: [0, 0, 0] }, 1, 0, this);
-    this.s = PropertyFactory.create(data.s || { k: [100, 100, 100] }, 1, 0.01, this);
-    this.o = PropertyFactory.create(data.o, 0, 0.01, this);
+  constructor(data: KeyFrames) {
+    const { create } = TransformFrames;
+
+    this.p = create(data.p || { k: [0, 0, 0] }, 1);
+    this.properties.push(this.p);
+
+    this.r = create(data.r || { k: 0 }, 0, degToRads);
+    this.properties.push(this.r);
+
+    this.a = create(data.a || { k: [0, 0, 0] }, 1);
+    this.properties.push(this.a);
+
+    this.s = create(data.s || { k: [100, 100, 100] }, 1, 0.01);
+    this.properties.push(this.s);
+
+    this.o = create(data.o, 0, 0.01);
+    this.properties.push(this.o);
 
     if (!this.properties.length) {
-      this.update(initialDefaultFrame, true);
+      this.update(initialDefaultFrame);
     }
   }
 
-  /**
-   * get transform
-   * @param {number} frameNum frameNum
-   */
-  update(frameNum) {
-    if (frameNum === this.frameId) {
-      return;
+  update(frameNum: number) {
+    const len = this.properties.length;
+
+    for (let i = 0; i < len; i += 1) {
+      this.properties[i].update(frameNum);
     }
 
-    this.updateProperties(frameNum);
-
-    if (this.autoOriented) {
-      this.updateOrientation();
+    if (this.autoOrient) {
+      // TODO
     }
-
-    this.frameId = frameNum;
-  }
-
-  /**
-   *
-   */
-  updateOrientation() {
-    let v1 = defaultVector;
-    let v2 = defaultVector;
-
-    const frameRate = this.elem.frameRate;
-    if (this.p && this.p.keyframes && this.p.getValueAtTime) {
-      if (this.p._caching.lastFrame <= this.p.keyframes[0].t) {
-        v1 = this.p.getValueAtTime((this.p.keyframes[0].t + 0.01) / frameRate, 0);
-        v2 = this.p.getValueAtTime(this.p.keyframes[0].t / frameRate, 0);
-      } else if (this.p._caching.lastFrame >= this.p.keyframes[this.p.keyframes.length - 1].t) {
-        v1 = this.p.getValueAtTime((this.p.keyframes[this.p.keyframes.length - 1].t / frameRate), 0);
-        v2 = this.p.getValueAtTime((this.p.keyframes[this.p.keyframes.length - 1].t - 0.05) / frameRate, 0);
-      } else {
-        v1 = this.p.pv;
-        v2 = this.p.getValueAtTime((this.p._caching.lastFrame - 0.01) / frameRate, this.p.offsetTime);
-      }
-    } else if (this.px && this.px.keyframes && this.py.keyframes && this.px.getValueAtTime && this.py.getValueAtTime) {
-      v1 = [];
-      v2 = [];
-      let px = this.px; let py = this.py;
-      if (px._caching.lastFrame+px.offsetTime <= px.keyframes[0].t) {
-        v1[0] = px.getValueAtTime((px.keyframes[0].t + 0.01) / frameRate, 0);
-        v1[1] = py.getValueAtTime((py.keyframes[0].t + 0.01) / frameRate, 0);
-        v2[0] = px.getValueAtTime((px.keyframes[0].t) / frameRate, 0);
-        v2[1] = py.getValueAtTime((py.keyframes[0].t) / frameRate, 0);
-      } else if (px._caching.lastFrame+px.offsetTime >= px.keyframes[px.keyframes.length - 1].t) {
-        v1[0] = px.getValueAtTime((px.keyframes[px.keyframes.length - 1].t / frameRate), 0);
-        v1[1] = py.getValueAtTime((py.keyframes[py.keyframes.length - 1].t / frameRate), 0);
-        v2[0] = px.getValueAtTime((px.keyframes[px.keyframes.length - 1].t - 0.01) / frameRate, 0);
-        v2[1] = py.getValueAtTime((py.keyframes[py.keyframes.length - 1].t - 0.01) / frameRate, 0);
-      } else {
-        v1 = [px.pv, py.pv];
-        v2[0] = px.getValueAtTime((px._caching.lastFrame+px.offsetTime - 0.01) / frameRate, px.offsetTime);
-        v2[1] = py.getValueAtTime((py._caching.lastFrame+py.offsetTime - 0.01) / frameRate, py.offsetTime);
-      }
-    }
-    this.orientation = -Math.atan2(v1[1] - v2[1], v1[0] - v2[0]);
   }
 }
