+++ src/core/bez/BezierEaser.ts
