+++ src/core/bez/BezierEasing.ts
