+++ src/core/index.ts
@@ -1,6 +1,6 @@
-import CompLottieLayer from './CompLottieLayer';
-import SpriteLottieLayer from './SpriteLottieLayer';
-import Tools from './utils/tools';
+import CompLottieLayer from './layer/CompLottieLayer';
+import SpriteLottieLayer from './layer/SpriteLottieLayer';
+import Tools from './tools';
 
 export {
   CompLottieLayer,
