+++ src/core/property/KeyframedMultidimensionalProperty.ts
@@ -1,0 +1,43 @@
+
+import bez from '../bez/bez';
+import { initialDefaultFrame as initFrame } from '../contant';
+import BaseProperty from './BaseProperty';
+
+/**
+ * keyframed multidimensional value property
+ */
+export default class KeyframedMultidimensionalProperty extends BaseProperty {
+  constructor(data, mult) {
+    super(data, mult);
+    this.propType = 'multidimensional';
+    let i; let len = data.k.length;
+    let s; let e; let to; let ti;
+    for (i = 0; i < len - 1; i += 1) {
+      if (data.k[i].to && data.k[i].s && data.k[i + 1] && data.k[i + 1].s) {
+        s = data.k[i].s;
+        e = data.k[i + 1].s;
+        to = data.k[i].to;
+        ti = data.k[i].ti;
+        // console.log('bez', bez)
+        if ((s.length === 2 && !(s[0] === e[0] && s[1] === e[1]) && bez.pointOnLine2D(s[0], s[1], e[0], e[1], s[0] + to[0], s[1] + to[1]) && bez.pointOnLine2D(s[0], s[1], e[0], e[1], e[0] + ti[0], e[1] + ti[1])) || (s.length === 3 && !(s[0] === e[0] && s[1] === e[1] && s[2] === e[2]) && bez.pointOnLine3D(s[0], s[1], s[2], e[0], e[1], e[2], s[0] + to[0], s[1] + to[1], s[2] + to[2]) && bez.pointOnLine3D(s[0], s[1], s[2], e[0], e[1], e[2], e[0] + ti[0], e[1] + ti[1], e[2] + ti[2]))) {
+          data.k[i].to = null;
+          data.k[i].ti = null;
+        }
+        if (s[0] === e[0] && s[1] === e[1] && to[0] === 0 && to[1] === 0 && ti[0] === 0 && ti[1] === 0) {
+          if (s.length === 2 || (s[2] === e[2] && to[2] === 0 && ti[2] === 0)) {
+            data.k[i].to = null;
+            data.k[i].ti = null;
+          }
+        }
+      }
+    }
+    this.kf = true;
+    let arrLen = data.k[0].s.length;
+    this.v = new Float32Array(arrLen);
+    for (i = 0; i < arrLen; i += 1) {
+      this.v[i] = initFrame * this.mult;
+    }
+    this._caching = { lastFrame: initFrame, lastIndex: 0, value: new Float32Array(arrLen) };
+
+  }
+}
