+++ src/core/tools.ts
