--- src/core/utils/pooling/BezierLengthPool.js
@@ -1,18 +1,0 @@
-import { createTypedArray } from '../helpers/arrays';
-import PoolFactory from './PoolFactory';
-const defaultCurveSegments = 200;
-/**
- * a
- * @private
- * @return {*}
- */
-function create() {
-  return {
-    addedLength: 0,
-    percents: createTypedArray('float32', defaultCurveSegments),
-    lengths: createTypedArray('float32', defaultCurveSegments),
-  };
-}
-const BezierLengthPool = PoolFactory(8, create);
-
-export default BezierLengthPool;
