--- src/core/utils/pooling/SegmentsLengthPool.js
@@ -1,30 +1,0 @@
-import BezierLengthPool from './BezierLengthPool';
-import PoolFactory from './PoolFactory';
-
-/**
- * @private
- * @return {*}
- */
-function create() {
-  return {
-    lengths: [],
-    totalLength: 0,
-  };
-}
-
-/**
- * a
- * @private
- * @param {*} element a
- */
-function release(element) {
-  const len = element.lengths.length;
-  for (let i = 0; i < len; i += 1) {
-    BezierLengthPool.release(element.lengths[i]);
-  }
-  element.lengths.length = 0;
-}
-
-const SegmentsLengthPool = PoolFactory(8, create, release);
-
-export default SegmentsLengthPool;
